Move Example functions to example_test.go
Fixes devlights/gomy#117
