+++ chans/concat_test.go
@@ -1,40 +1,11 @@
 package chans_test
 
 import (
-	"context"
-	"fmt"
 	"reflect"
 	"testing"
-	"time"
 
 	"github.com/devlights/gomy/chans"
 )
-
-func ExampleConcat() {
-	var (
-		rootCtx          = context.Background()
-		mainCtx, mainCxl = context.WithCancel(rootCtx)
-		procCtx, procCxl = context.WithTimeout(mainCtx, 50*time.Millisecond)
-	)
-
-	defer mainCxl()
-	defer procCxl()
-
-	nums1 := chans.Generator(procCtx.Done(), 1, 2, 3)
-	nums2 := chans.Generator(procCtx.Done(), 4, 5, 6)
-
-	for v := range chans.Concat(procCtx.Done(), nums1, nums2) {
-		fmt.Println(v)
-	}
-
-	// Output:
-	// 1
-	// 2
-	// 3
-	// 4
-	// 5
-	// 6
-}
 
 func TestConcat(t *testing.T) {
 	type (
