+++ chans/ordone_test.go
@@ -2,40 +2,11 @@ package chans_test
 
 import (
 	"context"
-	"fmt"
 	"testing"
 	"time"
 
 	"github.com/devlights/gomy/chans"
 )
-
-func ExampleOrDone() {
-	var (
-		rootCtx          = context.Background()
-		mainCtx, mainCxl = context.WithCancel(rootCtx)
-		procCtx, procCxl = context.WithTimeout(mainCtx, 1*time.Minute)
-		genCtx, genCxl   = context.WithCancel(mainCtx)
-	)
-
-	defer mainCxl()
-	defer procCxl()
-	defer genCxl()
-
-	inCh := chans.Generator(genCtx.Done(), "h", "e", "l", "l", "o")
-
-	var result []interface{}
-	for v := range chans.OrDone(procCtx.Done(), inCh) {
-		func() {
-			defer procCxl()
-			result = append(result, v)
-		}()
-	}
-
-	fmt.Printf("len(result) <= 2: %v", len(result) <= 2)
-
-	// Output:
-	// len(result) <= 2: true
-}
 
 func TestOrDone(t *testing.T) {
 	data := make([]interface{}, 0, 200)
