+++ chans/whenall_test.go
@@ -1,37 +1,11 @@
 package chans_test
 
 import (
-	"fmt"
 	"testing"
 	"time"
 
 	"github.com/devlights/gomy/chans"
 )
-
-func ExampleWhenAll() {
-	fn := func(tlimit time.Duration) <-chan struct{} {
-		done := make(chan struct{})
-		go func() {
-			defer close(done)
-			time.Sleep(tlimit)
-		}()
-
-		return done
-	}
-
-	done1 := fn(100 * time.Millisecond)
-	done2 := fn(200 * time.Millisecond)
-	done3 := fn(300 * time.Millisecond)
-
-	start := time.Now()
-	<-chans.WhenAll(done1, done2, done3)
-	elapsed := time.Since(start)
-
-	fmt.Printf("elapsed: about 300msec ==> %v\n", elapsed >= 299*time.Millisecond)
-
-	// Output:
-	// elapsed: about 300msec ==> true
-}
 
 func TestWhenAll(t *testing.T) {
 	type (
