+++ chans/whenany_test.go
@@ -1,37 +1,11 @@
 package chans_test
 
 import (
-	"fmt"
 	"testing"
 	"time"
 
 	"github.com/devlights/gomy/chans"
 )
-
-func ExampleWhenAny() {
-	fn := func(tlimit time.Duration) <-chan struct{} {
-		done := make(chan struct{})
-		go func() {
-			defer close(done)
-			time.Sleep(tlimit)
-		}()
-
-		return done
-	}
-
-	done1 := fn(100 * time.Millisecond)
-	done2 := fn(200 * time.Millisecond)
-	done3 := fn(300 * time.Millisecond)
-
-	start := time.Now()
-	<-chans.WhenAny(done1, done2, done3)
-	elapsed := time.Since(start)
-
-	fmt.Printf("elapsed: about 100msec ==> %v\n", elapsed <= 110*time.Millisecond)
-
-	// Output:
-	// elapsed: about 100msec ==> true
-}
 
 func TestWhenAny(t *testing.T) {
 	type (
